Update style.css
text-aligned the nav bar centered, so that when we hover over them with an additional option (contact page) it will be centered on the button. Again, not sure if this was removed intentionally, and if so that is totally fine. Just toying with it to see if it'll change the formatting